+++ lib/src/style/line_style_logger_formatter.dart
@@ -8,7 +8,7 @@ class LineStyleLogger implements StyleSource {
 
   @override
   String formater(LogEntity details, HybridSettings settings) {
-    final header = _formatHeader(details.header);
+    final header = _formatHeader(details.header, settings);
     final message = _formatMessage(details.message, settings);
     final logTypeHandlers = _getLogTypeHandlers(details, settings);
     final formattedLog = _formatLogLines(
@@ -21,8 +21,18 @@ class LineStyleLogger implements StyleSource {
     return _applyColorAndJoin(formattedLog, details.color.write);
   }
 
-  String _formatHeader(String? header) {
-    return header == null ? '' : '[$header]';
+  String _formatHeader(String? header, HybridSettings settings) {
+    if (header == null) {
+      return '';
+    }
+
+    String? trimmedHeader;
+
+    if (header.length > settings.maxLineWidth) {
+      trimmedHeader = settings.limitHeaderLength ? '${header.substring(0, settings.maxLineWidth - 4)}...' : header;
+    }
+
+    return '[${trimmedHeader ?? header}]';
   }
 
   String _formatMessage(dynamic message, HybridSettings settings) {
@@ -34,23 +44,18 @@ class LineStyleLogger implements StyleSource {
     return trimmedMessage;
   }
 
-  Map<String, String> _getLogTypeHandlers(
-      LogEntity details, HybridSettings settings) {
+  Map<String, String> _getLogTypeHandlers(LogEntity details, HybridSettings settings) {
     return {
-      'stackTrace':
-          details.type == LogTypeEntity.stacktrace && details.stack != null
-              ? _formatStackTrace(details.stack!, settings)
-              : '',
-      'httpError':
-          details.type == LogTypeEntity.httpError && details.httpError != null
-              ? _formatHttpError(details.httpError!, settings)
-              : '',
-      'httpResponse': details.type == LogTypeEntity.httpResponse &&
-              details.httpResponse != null
+      'stackTrace': details.type == LogTypeEntity.stacktrace && details.stack != null
+          ? _formatStackTrace(details.stack!, settings)
+          : '',
+      'httpError': details.type == LogTypeEntity.httpError && details.httpError != null
+          ? _formatHttpError(details.httpError!, settings)
+          : '',
+      'httpResponse': details.type == LogTypeEntity.httpResponse && details.httpResponse != null
           ? _formatHttpResponse(details.httpResponse!, settings)
           : '',
-      'httpRequest': details.type == LogTypeEntity.httpRequest &&
-              details.httpRequest != null
+      'httpRequest': details.type == LogTypeEntity.httpRequest && details.httpRequest != null
           ? _formatHttpRequest(details.httpRequest!, settings)
           : '',
     };
@@ -86,9 +91,7 @@ class LineStyleLogger implements StyleSource {
   }
 
   List<String> _selectLogLines(Map<String, String> logTypeHandlers) {
-    return logTypeHandlers.values
-        .expand((log) => _indentLines(log.split('\n')))
-        .toList();
+    return logTypeHandlers.values.expand((log) => _indentLines(log.split('\n'))).toList();
   }
 
   List<String> _indentLines(List<String> lines) {
@@ -102,17 +105,15 @@ class LineStyleLogger implements StyleSource {
     const String reset = '\x1B[0m';
     const String gray = '\x1B[38;5;247m';
     return lines.map((line) {
-      final String prefixedLine = '$gray[MT-LOG]$reset $line';
-      final String coloredLine =
-          prefixedLine.replaceFirst(line, colorWriter(line));
+      final String prefixedLine = '$gray[HL-LOG]$reset $line';
+      final String coloredLine = prefixedLine.replaceFirst(line, colorWriter(line));
       return coloredLine;
     }).join('\n');
   }
 
   String _formatStackTrace(StackTrace stack, HybridSettings settings) {
     final stackEntity = _parseTrace(stack);
-    final stackString =
-        [stackEntity.fileName, '${stackEntity.functionName}()'].join('\n');
+    final stackString = [stackEntity.fileName, '${stackEntity.functionName}()'].join('\n');
 
     if (settings.maxLogLength == null) return stackString;
 
@@ -134,10 +135,8 @@ class LineStyleLogger implements StyleSource {
     return trimmedMessage;
   }
 
-  String _formatHttpResponse(
-      HybridHttpResponse response, HybridSettings settings) {
-    final String responseString =
-        " => RESPONSE STATUS CODE: ${response.statusCode}\n"
+  String _formatHttpResponse(HybridHttpResponse response, HybridSettings settings) {
+    final String responseString = " => RESPONSE STATUS CODE: ${response.statusCode}\n"
         " => RESPONSE STATUS MESSAGE: ${response.statusMessage} \n"
         " => RESPONSE DATA: ${response.data}\n"
         " => MS: ${response.ms}";
@@ -150,8 +149,7 @@ class LineStyleLogger implements StyleSource {
     return trimmedMessage;
   }
 
-  String _formatHttpRequest(
-      HybridHttpRequest request, HybridSettings settings) {
+  String _formatHttpRequest(HybridHttpRequest request, HybridSettings settings) {
     final String requestString = " => BASE URL: ${request.baseUrl}\n"
         " => PATH: ${request.path}\n"
         " => DATA: ${request.data}\n"